extract huis schema and preload hook
