+++ src/Mappools.tsx
@@ -1,17 +1,18 @@
-import { motion } from "framer-motion";
-import { sectionVariants, getAnimations } from "./animations";
-import type { AnimTypes } from "./animations";
 import clsx from "clsx";
+import { motion } from "framer-motion";
+import type { AnimTypes } from "./animations";
+import { getAnimations, sectionVariants } from "./animations";
 import { Casters } from "./components/Casters";
 import { Chat } from "./components/Chat";
+import { FooterContent } from "./components/FooterContent";
+import { HeaderContent } from "./components/HeaderContent";
 import { Logo } from "./components/Logo";
 import { MainContent } from "./components/MainContent";
-import { HeaderContent } from "./components/HeaderContent";
-import { FooterContent } from "./components/FooterContent";
 import { PlayerInfo } from "./components/PlayerInfo";
 import { StageInfo } from "./components/StageInfo";
-import { useMappoolQuery, type Beatmap } from "./state/huis";
 import { useSettings } from "./state/dashboard";
+import type { Beatmap } from "./schemas/huis";
+import { useMappoolQuery } from "./state/huis";
 
 function ModBracket(
   beatmaps: Beatmap[],
