+++ src/Screens.tsx
@@ -6,19 +6,10 @@ import { StartScreen } from "./Startscreen";
 import { VersusScreen } from "./Versus";
 import { WinnerScreen } from "./Winner";
 import { useSettings } from "./state/dashboard";
-import {
-  useMappoolQuery,
-  useMatchesQuery,
-  useScheduleQuery,
-  useTournamentQuery,
-} from "./state/huis";
+import { usePreload } from "./state/preload";
 
 export function Screens() {
-  // prefetching
-  useMatchesQuery();
-  useMappoolQuery();
-  useTournamentQuery();
-  useScheduleQuery();
+  usePreload();
 
   const [settings] = useSettings();
   const activeScreen = settings.activeScreen;
