+++ src/state/preload.ts
@@ -1,0 +1,13 @@
+import {
+  useMatchesQuery,
+  useMappoolQuery,
+  useTournamentQuery,
+  useScheduleQuery,
+} from "./huis";
+
+export function usePreload() {
+  useMatchesQuery();
+  useMappoolQuery();
+  useTournamentQuery();
+  useScheduleQuery();
+}
